+++ shared/Module Entities/Group.ts
@@ -1,6 +1,5 @@
 import { Module } from './Module'
 import { ModuleEntity } from './ModuleEntity'
-import { v4 as UUIDV4 } from 'uuid'
 
 /** Represents a Group in a Module */
 
@@ -12,9 +11,11 @@ export class Group extends ModuleEntity {
   /**
    * Initiaizes an instance of `Group`
    * @param name The name of the group
+   * @param moduleUUID The UUID of the module
    */
-  constructor(name: string = 'Unnamed Group') {
+  constructor(name: string = 'Unnamed Group', moduleUUID: string) {
     let slug = Module.getSlugFromValue(`group-${Module.getSlugFromValue(name)}`)
-    super(name, UUIDV4(), slug)
+    
+    super(name, moduleUUID, slug)
   }
 }