+++ shared/Module Entities/ModuleEntity.ts
@@ -1,4 +1,5 @@
 import { Module } from "./Module"
+import { v5 as UUIDV5 } from 'uuid'
 
 /** Represents referencable entities in a module */
 export abstract class ModuleEntity {
@@ -9,11 +10,11 @@ export abstract class ModuleEntity {
   /**
    * Initializes an instance of `ModuleEntity`
    * @param name The name of the entity
-   * @param id The UUID of the entity
+   * @param moduleUUID The UUID of the module
    */
   constructor(
     readonly name: string,
-    readonly id: string,
+    readonly moduleUUID: string,
     slug: string | undefined = undefined
   ) {
     // If a slug is manually specified, make sure it doesn't exist.
@@ -28,8 +29,15 @@ export abstract class ModuleEntity {
       }
     }
 
+    // Once the slug has been determined, add to the list of existing slugs
+    // so there can't be overlap
     this.slug = slug ?? Module.getSlugFromValue(name)
     Module.existingSlugs.push(this.slug)
+
+    // Derive UUID from slug using module ID as namespace. This
+    // makes IDs deterministic for a given module (so repeated
+    // packing results in same IDs).
+    this.id = UUIDV5(this.slug, moduleUUID) 
   }
 
   // ---------------------------------------------------------------
@@ -41,4 +49,7 @@ export abstract class ModuleEntity {
 
   /** The slug for the entity */
   slug: string
+
+  /** The ID of the entity (UUIDV5) */
+  id: string
 }