+++ shared/Module Entities/Page.ts
@@ -1,5 +1,4 @@
 import { ModuleEntity } from './ModuleEntity'
-import { v4 as UUIDV4 } from "uuid"
 
 /** Represents a Page in a Module */
 
@@ -11,12 +10,16 @@ export class Page extends ModuleEntity {
   /**
    * Initiaizes an instance of `Page`
    * @param name The name of the page
+   * @param moduleUUID The UUID of the module
+   * @param slug A manually specified slug (optional - will be auto-generated if undefined)
    */
   constructor(
-    name: string = "Unnamed Page",
-    slug: string | undefined = undefined
-  ) {
-    super(name, UUIDV4(), slug)
+    name: string,
+    moduleUUID: string,
+    slug: string | undefined = undefined,
+    
+  ) {    
+    super(name, moduleUUID, slug)
   }
 
   // ---------------------------------------------------------------